fix: BundleDetailResponse writer 타입 Member -> WriterResponse
