+++ server/api/src/main/kotlin/com/kw/api/domain/bundle/dto/response/BundleDetailResponse.kt
@@ -1,5 +1,6 @@
 package com.kw.api.domain.bundle.dto.response
 
+import com.kw.api.domain.member.dto.response.SnsResponse
 import com.kw.api.domain.question.dto.response.QuestionResponse
 import com.kw.api.domain.tag.dto.response.TagResponse
 import com.kw.data.domain.bundle.Bundle
@@ -14,23 +15,52 @@ data class BundleDetailResponse(
     val tags: List<TagResponse>? = null,
     val questions: List<QuestionResponse>? = null,
     val originId: Long? = null,
-    val writer: Member?,
+    val writer: WriterResponse?,
     val createdAt: LocalDateTime? = null,
     val updatedAt: LocalDateTime? = null
 ) {
+
+
     companion object {
-        fun of(bundle: Bundle, questions: List<Question>, memberId: Long?): BundleDetailResponse {
+        fun of(bundle: Bundle, questions: List<Question>, requestMemberId: Long?): BundleDetailResponse {
             return BundleDetailResponse(
                 id = bundle.id!!,
                 name = bundle.name,
                 shareType = bundle.shareType.name,
                 tags = bundle.bundleTags.map { it.tag }.map { TagResponse.from(it) },
-                questions = questions.map { QuestionResponse.from(it, memberId) },
+                questions = questions.map { QuestionResponse.from(it, requestMemberId) },
                 originId = bundle.originId,
-                writer = bundle.member,
+                writer = bundle.member?.let { WriterResponse.from(it) },
                 createdAt = bundle.createdAt,
                 updatedAt = bundle.updatedAt
             )
+        }
+    }
+
+    data class WriterResponse(
+        val id: Long,
+        val nickname: String?,
+        val profileImage: String?,
+        val email: String,
+        val snsList: List<SnsResponse>,
+        val tags: List<TagResponse>,
+        val createdAt: LocalDateTime,
+        val updatedAt: LocalDateTime
+    ) {
+
+        companion object {
+            fun from(writer: Member): WriterResponse {
+                return WriterResponse(
+                    id = writer.id!!,
+                    nickname = writer.nickname,
+                    profileImage = writer.profileImage,
+                    email = writer.email,
+                    snsList = writer.snsList.map { SnsResponse.from(it) },
+                    tags = writer.memberTags.map { TagResponse.from(it.tag) },
+                    createdAt = writer.createdAt,
+                    updatedAt = writer.updatedAt
+                )
+            }
         }
     }
 }
